Revert ""fix" compile balrok outdoor"
This reverts commit 06a881da1c37939d0cb7c61f9bb67f02589038d2.

+++ src/game/Unit.cpp
@@ -6340,7 +6340,7 @@ bool Unit::HandleDummyAuraProc(Unit *pVictim, uint32 damage, Aura* triggeredByAu
                     basepoints[0] += damage * triggerAmount/100;
 
                     // limit absorb amount
-                    if (basepoints[0] > pVictim->getLevel()*125)
+                    if (uint32(basepoints[0]) > pVictim->getLevel()*125)
                         basepoints[0] = pVictim->getLevel()*125;
 
                     triggered_spell_id = 47753;
